// generated/src/app/api-models/diag-code-master-staging.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class DiagCodeMasterStaging {
  diagnosisCode: string;
  shortDescription: string;
  description: string;
  dxClass1: string;
  dxClass2: string;
  dxClass3: string;
  fromAge: number;
  toAge: number;
  patientGender: string;
  traumaFlag: string;
  caseManagement: string;
  procedureClass: string;
  mdcCode: string;
  editFlag: string;
  operProcedureFlag: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  drgDefaultBirthweight: number;
  effectiveDate: Date;
  termDate: Date;
}